% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{attributes}
\alias{attributes}
\title{Find attributes}
\usage{
attributes(x)
}
\arguments{
\item{x}{string}
}
\value{
string
}
\description{
find attributes in dataset, in the biomart database given a search string
}
\examples{
attributes("pfam")
}
